Mise en place d'un vrai moteur physique basique (AABBs sans rotation)

+++ ska_ai/AI/System/IARandomMovementSystem.cpp
@@ -32,7 +32,7 @@ void ska::IARandomMovementSystem::refresh(unsigned int) {
 						const auto& pcEmitterPtr = m_componentPossibleAccessor.get<PositionComponent>(iamc.emitter);
                         if (pcEmitterPtr != nullptr) {
                             auto& pcEmitter = *pcEmitterPtr;
-                            auto diffPoint = Point<int>(pcEmitter.x, pcEmitter.y) - Point<int>(pc.x, pc.y);
+                            auto diffPoint = Point<float>(pcEmitter.x, pcEmitter.y) - Point<float>(pc.x, pc.y);
                             auto horizontalDir = NumberUtils::maximum(NumberUtils::absolute(diffPoint.x), NumberUtils::absolute(diffPoint.y)) == NumberUtils::absolute(diffPoint.x);
                             iamc.direction = horizontalDir ? (sens ? 1 : 3) : (sens ? 0 : 2);
