+++ ska_draw/Graphic/System/CameraFollowSystem.cpp
@@ -12,7 +12,7 @@ void ska::CameraFollowSystem::refresh(unsigned int) {
 	if(mainEntityIt != processed.end()) {
 		auto entityId = *mainEntityIt;
 		auto& pc = m_componentAccessor.get<PositionComponent>(entityId);
-		m_follow = { pc.x, pc.y, 0, 0 };
+		m_follow = {static_cast<int>(pc.x), static_cast<int>(pc.y), 0, 0 };
 		m_pos = &m_follow;
 		focusOn(m_follow, &entityId);
 	}
